[storage] Fix race condition in backup
If multiple backups were requested at the same time when the backup was
older than the set maximum age for backups, then it would launch the
full backup process several times over the same file, leading to the
backup method finishing early save for the first, and the web server
serving the backup file while it is still open by sqlite.

This is simply solved by a lock. Slight refactorization of the Storage
data structure was also done for readability.